Move branch filter logic
By moving it up here, we make sure only jobs that should be built is added to the JOBS array in the first place.
This is a better way, since the skipped jobs won't occupy the build slots (since we take X number of jobs from the JOBS array for each batch, for a CONCURRENT_JOBS_COUNT of 10 - if 6 is skipped, only 4 jobs will be started).

Now, they are not added to the JOBS array, so the rest works as intended.